pyproject.toml
[project]
name = "jrdbdownloader"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "maoki", email = "michihito.aoki@gmail.com" }
]
dependencies = [
    "beautifulsoup4>=4.12.3",
    "python-dotenv>=1.0.1",
    "requests>=2.32.3",
    "tqdm>=4.66.4",
]
readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = [
    "pip>=24.0",
    "ruff>=0.4.9",
    "spyder-kernels==2.5.*",
    "jupyterlab>=4.2.2",
    "notebook>=7.2.1",
    "jupyterlab-lsp>=5.1.0",
    "python-lsp-server>=1.11.0",
    "python-lsp-ruff>=2.2.1",
    "jupyterlab-code-formatter>=2.2.1",
]

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/jrdbdownloader"]
